rename MetaValidationUtils -> NameValidator
