+++ molgenis-data-import/src/main/java/org/molgenis/data/importer/EmxMetaDataParser.java
@@ -10,7 +10,6 @@ import org.molgenis.data.i18n.model.I18nString;
 import org.molgenis.data.i18n.model.I18nStringFactory;
 import org.molgenis.data.i18n.model.Language;
 import org.molgenis.data.i18n.model.LanguageFactory;
-import org.molgenis.data.meta.MetaValidationUtils;
 import org.molgenis.data.meta.SystemEntityType;
 import org.molgenis.data.meta.model.*;
 import org.molgenis.data.meta.model.Package;
