+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/meta/AttributeValidator.java
@@ -18,7 +18,7 @@ import java.util.Objects;
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
 import static org.molgenis.AttributeType.*;
-import static org.molgenis.data.meta.MetaValidationUtils.validateName;
+import static org.molgenis.data.meta.NameValidator.validateName;
 import static org.molgenis.data.meta.model.AttributeMetadata.ATTRIBUTE_META_DATA;
 import static org.molgenis.data.meta.model.EntityTypeMetadata.ENTITY_TYPE_META_DATA;
 import static org.molgenis.data.meta.model.PackageMetadata.PACKAGE;
